pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["alexander <alexander12110@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
loguru = "^0.7.2"
asgi-correlation-id = "^4.3.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
